+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## Version 1.0.57
+
+Fixed an issue with combat turns in v9.  Refrences to the token involved have changed.
+
+Clarified what's happening if you try and convert lootables but no loot sheet is selected.
+
+## Version 1.0.56
+
+When the saving throw request is closed, informing Active Tiles that the state information is no longer needed.
+
+Moved the token indicators higher in the z-order so they appear above the bars.
+
+Added the option to change the global movement state if no entity is selected.
+
+Added flavor text to the request roll via Active Tiles.
+
 ## Version 1.0.55
 
 Added option to return the chat message created when prompting to assign xp